[feat]: 넘버 피커 값 받아오기

+++ feature/onboarding/src/main/java/com/hmh/hamyeonham/feature/onboarding/model/OnboardingAnswer.kt
@@ -9,6 +9,6 @@ data class OnboardingAnswer(
 ) {
     data class App(
         val appCode: String = "",
-        val goalTime: Int = -1,
+        val goalTime: Long = -1,
     )
 }
